requirements-visual.txt
altair
jupyter
pandas
pillow
pyvirtualdisplay
torchviz
